+++ spec/acceptance/cmd_runuser_spec.rb
@@ -18,8 +18,9 @@ describe 'initscript context' do
       ->
 
       initscript::service { 'evilcmd':
-        cmd => 'sleep 666m',
+        cmd      => 'sleep 666m',
         run_user => 'evil',
+        debug    => '/evilcmd.log',
       }
 
       ->
@@ -34,6 +35,14 @@ describe 'initscript context' do
       # Run it twice and test for idempotency
       expect(apply_manifest(pp).exit_code).to_not eq(1)
       expect(apply_manifest(pp).exit_code).to eq(0)
+    end
+
+    it "debug service" do
+     expect(shell("cat /etc/init.d/evilcmd").exit_code).to be_zero
+    end
+
+    it "debug log" do
+     expect(shell("cat /evilcmd.log").exit_code).to be_zero
     end
 
     it "process started" do