+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 0.1.8
+
+* bugfix **initscript::service**
+* added debug options to **initscript::service**
+
+## 0.1.7
+
+* improved init script
+
 ## 0.1.6
 
 * added **run_user** to **initscript::service**